Updates
Corrige ordem das variáveis em clientes.py, visualização de cadastros em cadastro_clientes.py e transfere menu para main.py
