+++ farmacia/cadastro_cliente.py
@@ -16,10 +16,12 @@ class CadastroCliente:
         else:
             raise ValueError('CPF já cadastrado')
             
-    def visualizar_cadastro(self, cpf) -> None:
-        for cpf in self.cadastrados:
-            if cpf == self.cadastrados[cpf]:
-                print(self.cadastrados[cpf])
+    def visualizar_cadastro(self, cpf):
+        for cadastro in self.cadastrados:
+            if cpf == cadastro:
+                print(self.cadastrados[cadastro])
+            else:
+                print('CPF não cadastrado')
             
     def alterar_cadastro(self, cpf):
         if cpf not in self.cadastrados:
@@ -45,32 +47,3 @@ class CadastroCliente:
             self.cadastrados.pop(cpf)
             
         self.cadastrados[cliente.cpf] = cliente
-        
-    def run(self):
-        
-        print('Menu\n1 - Novo cadastro\n2 - Procurar cadastro\n3 - Alterar cadastro\n0 - Sair')
-        acao_menu = int(input('O que deseja fazer? '))
-        
-        while acao_menu != 0:
-            
-            if acao_menu == 1:
-                print('===Novo cadastro===\n')
-                self.cadastrar_novo()
-                print('\n===Fim do cadastro===')
-                
-            elif acao_menu == 2:
-                print('===Procurar cadastrado===\n')
-                cpf_a_procurar = input('Digite o CPF: ')
-                self.visualizar_cadastro(cpf_a_procurar)
-                print('\n===Fim da visualização===')
-            
-            elif acao_menu == 3:
-                print('===Alterar cadastros===\n')
-                cpf_a_alterar = input('Digite o CPF a ser alterado: ')
-                self.alterar_cadastro(cpf_a_alterar)
-                print('\n===Fim da alteração===')
-            
-            print('Menu\n1 - Novo cadastro\n2 - Visualizar cadastros\n3 - Alterar cadastro\n0 - Sair')
-            acao_menu = int(input("O que deseja fazer? "))
-        
-        print('===Fim===')
